+++ sdk/tables/data-tables/test/internal/serialization.spec.ts
@@ -5,6 +5,7 @@ import { assert } from "chai";
 
 import { Edm } from "../../src";
 import { serialize, deserialize } from "../../src/serialization";
+import { isNode8 } from "@azure/test-utils";
 
 interface Entity {
   strProp?: string;
@@ -163,7 +164,10 @@ describe("Deserializer", () => {
     assert.strictEqual(deserialized.int32Prop, int32Value);
   });
 
-  it("should deserialize an Int64 value to bigint", () => {
+  it("should deserialize an Int64 value to bigint", function(this: Mocha.Context) {
+    if (isNode8) {
+      this.skip();
+    }
     const int64Value = "12345678910";
     const deserialized = deserialize({
       int64ObjProp: int64Value,